feat: Add step information for executions of multiple steps at a time

+++ src/dirac_cwl/commands/workflow_commons.py
@@ -21,7 +21,9 @@ from DIRAC.TransformationSystem.Client.FileReport import FileReport
 from DIRAC.WorkloadManagementSystem.Client.JobReport import JobReport
 from LHCbDIRAC.BookkeepingSystem.Client.BookkeepingClient import BookkeepingClient
 from LHCbDIRAC.Core.Utilities.XMLSummaries import XMLSummary
-from pydantic import BaseModel, ConfigDict, PrivateAttr
+from pydantic import BaseModel, ConfigDict, Field, PrivateAttr
+
+logger = logging.getLogger(__name__)
 
 
 class StepStatus(str, Enum):
@@ -29,6 +31,83 @@ class StepStatus(str, Enum):
 
     Done = "Done"
     Failed = "Failed"
+
+
+class Step(BaseModel):
+    """Execution step information."""
+
+    id: str
+    name: str
+    number: int
+
+    executable: str = "gaudirun.py"
+
+    application_name: Optional[str] = "Unknown"
+    cleaned_application_name: str = ""
+    application_version: str = "Unknown"
+    application_log: str = ""
+    application_type: str = ""
+
+    event_type: str = ""
+    number_of_events: int = 0
+    event_timeout: Optional[int] = None
+
+    extra_packages: Optional[str] = ""
+    proc_pass: str = ""
+    bk_id: str = ""
+    multicore: bool = False
+    mc_tck: str = ""
+    system_config: str = ""
+
+    dddb_tag: str = ""
+    conddb_tag: str = ""
+    dq_tag: str = ""
+
+    inputs: list[str] = []
+    outputs: list[dict[str, Any]] = []
+
+    input_data_type: str = ""
+
+    options_file: str = ""
+    options_line: str = ""
+    extra_options_line: str = ""
+    options_format: str = ""
+
+    size: dict = {}
+    md5: dict = {}
+    guid: dict = {}
+
+    start_time: Optional[float] = None
+    start_stats: Optional[tuple] = None
+
+    # To be built if certain conditions are met
+    # > If (wf_c.production_id && wf_c.job_id && self.name && self.inputs)
+    output_file_prefix: str = ""
+    xml_summary_path: str = ""
+    histo_name: str = "Hist.root"
+
+    # Private Attributes
+    _xf_o: Optional[XMLSummary] = PrivateAttr(default=None)
+
+    def __init__(self, **data):
+        """StepCommons constructor."""
+        super().__init__(**data)
+
+        if self.application_name:
+            self.cleaned_application_name = self.application_name.replace("/", "")
+
+        if self.xml_summary_path:
+            self._xf_o = XMLSummary(self.xml_summary_path)
+
+    @property
+    def xf_o(self) -> XMLSummary:
+        """Xml Summary getter."""
+        return self._xf_o
+
+    @xf_o.setter
+    def xf_o(self, value: XMLSummary) -> None:
+        """Xml Summary getter."""
+        self._xf_o = value
 
 
 class WorkflowCommons(BaseModel):
@@ -40,112 +119,79 @@ class WorkflowCommons(BaseModel):
     production_id: str
     prod_job_id: str
 
-    inputs: list[str]
-    outputs: list[dict[str, Any]]
-
-    number_of_events: int
-    event_type: str
-
-    step_id: str
-    step_number: int
+    inputs: list[str] = []
+    outputs: list[dict[str, Any]] = []
 
     config_version: str
     config_name: str
 
-    # Optional values
-    executable: str = "gaudirun.py"
-    application_name: str = "Unknown"
-    application_version: str = "Unknown"
+    steps: list[Step] = []
 
+    # Optional values
     production_output_data: list[str] = []
     output_data_file_mask: str = ""
     output_data_type: str = ""
-    output_data_step: str = ""
-    output_SEs: dict[str, str] = {}
+    output_SEs: dict[str, list[str]] = {}  # output -> SE list
     output_mode: str = ""
+    output_data_step: str = ""
 
     log_target_path: str = ""
     log_file_path: str = ""
     log_lfn_path: str = ""
     log_dir: str = ""
 
-    application_log: str = ""
-    application_type: str = ""
-    cleaned_application_name: str = ""
-
     number_of_processors: int = 1
-    max_number_of_processors: int = 0
+    max_number_of_processors: Optional[int] = None
 
-    step_proc_pass: str = ""
     run_number: str = "Unknown"
     sim_description: str = "NoSimConditions"
-
-    file_descendents: list[str] = []
-    file_size_map: dict[str, str] = {}
-    file_md5_map: dict[str, str] = {}
-    file_guid_map: dict[str, str] = {}
 
     bookkeeping_lfns: list[str] = []
     prod_output_lfns: list[str] = []
 
+    file_descendents: list[str] = []
     file_report_files_dict: dict = {}
     accounting_registers: list = []
-    xml_summary_path: str = ""
+    xml_summary_paths: dict[str, str] = {}
     request_dict: dict = {}
     files_report_files_dict: dict = {}
 
-    bk_step_id: Optional[int] = None
-    start_stats: Optional[tuple] = None
-    start_time: Optional[float] = None
+    site_name: str = Field(default_factory=siteName)
+    multicore: bool = False
 
-    site_name: str = ""
     step_status: StepStatus = StepStatus.Done
 
     # Private attributes
     _request = PrivateAttr(default=None)
     _failover_request = PrivateAttr(default=None)
     _job_report = PrivateAttr(default=None)
-    _file_report = PrivateAttr(default=None)
-    _data_manager = PrivateAttr(default=None)
-    _bk_client = PrivateAttr(default=None)
-    _dsc = PrivateAttr(default=None)
-    _xf_o = PrivateAttr(default=None)
+    _file_report = PrivateAttr(default_factory=FileReport)
+    _data_manager = PrivateAttr(default_factory=DataManager)
+    _bk_client = PrivateAttr(default_factory=BookkeepingClient)
+    _dsc = PrivateAttr(default_factory=DataStoreClient)
 
     _logger = PrivateAttr(default=logging.getLogger(__name__))
 
-    model_config = ConfigDict(
-        validate_assignment=True,
-    )
+    model_config = ConfigDict(validate_assignment=True, extra="forbid")
 
     def __init__(self, **data):
         """WorkflowCommons constructor."""
         super().__init__(**data)
 
-        self.cleaned_application_name = self.application_name.replace("/", "")
-        self.site_name = siteName()
-
         self._request = Request(fromDict=self.request_dict)
-
         self._failover_request = FailoverTransfer(self.request)
 
-        self._job_report = JobReport(self.job_id)
-
-        self._file_report = FileReport()
+        self._job_report = JobReport(self.job_id, source="WorkflowCommons")
         self._file_report.statusDict = self.file_report_files_dict
-
-        self._data_manager = DataManager()
-        self._bk_client = BookkeepingClient()
-        self._dsc = DataStoreClient()
-
-        if self.xml_summary_path:
-            self._xf_o = XMLSummary(self.xml_summary_path)
 
     def save(self, job_path: os.PathLike, failed: bool = False) -> None:
         """Update the workflow_commons file to accomodate for the new values."""
+        logger.info("Saving workflow commons json file")
         wf_path = Path(job_path).joinpath("workflow_commons.json")
         wf_backup = Path(job_path).joinpath("workflow_commons.json.back")
 
-        shutil.move(wf_path, wf_backup)
+        if os.path.exists(wf_path):
+            shutil.move(wf_path, wf_backup)
 
         if failed:
             self.step_status = StepStatus.Failed
@@ -157,10 +203,11 @@ class WorkflowCommons(BaseModel):
             wf_dict = self.model_dump(mode="json")
             with open(wf_path, "w", encoding="utf-8") as f:
                 json.dump(wf_dict, f)
-        except Exception:
+        except Exception as e:
+            self._logger.exception("Failed to save the workflows commons in a file", exc_info=e)
             raise
         finally:
-            wf_backup.unlink()
+            wf_backup.unlink(missing_ok=True)
 
     @classmethod
     def load(cls, job_path: os.PathLike) -> WorkflowCommons:
@@ -216,16 +263,6 @@ class WorkflowCommons(BaseModel):
     def dsc(self) -> DataStoreClient:
         """DataStoreClient property getter."""
         return self._dsc
-
-    @property
-    def xf_o(self) -> XMLSummary:
-        """XMLSummary property getter."""
-        return self._xf_o
-
-    @xf_o.setter
-    def xf_o(self, xf_o: XMLSummary) -> None:
-        """XMLSummary property getter."""
-        self._xf_o = xf_o
 
     def generateFailoverFile(self):
         """Create a request.json file."""
